Move invites views, templates forms and emails to its own app
