+++ meinberlin/apps/projectinvites/emails.py
@@ -2,14 +2,14 @@ from meinberlin.apps.contrib.emails import Email
 
 
 class InviteParticipantEmail(Email):
-    template_name = 'meinberlin_projects/emails/invite_participant'
+    template_name = 'meinberlin_projectinvites/emails/invite_participant'
 
     def get_receivers(self):
         return [self.object.email]
 
 
 class InviteModeratorEmail(Email):
-    template_name = 'meinberlin_projects/emails/invite_moderator'
+    template_name = 'meinberlin_projectinvites/emails/invite_moderator'
 
     def get_receivers(self):
         return [self.object.email]
